Bugfix/adminstudentviewfunctionality (#117)
* add flags to Student class. Use flags to know when student profile is complete for review

* added UT for Student class

+++ src/main/java/GAMS/controllers/StudentController.java
@@ -2,11 +2,11 @@ package GAMS.controllers;
 
 import GAMS.Crudrepository.EndUserRepo;
 import GAMS.Crudrepository.FieldOfResearchRepo;
+import GAMS.Crudrepository.StudentRepo;
 import GAMS.entity.EndUser;
 import GAMS.entity.FieldOfResearch;
 import GAMS.entity.Student;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
@@ -28,6 +28,9 @@ public class StudentController {
     @Autowired
     FieldOfResearchRepo researchRepository;
 
+    @Autowired
+    StudentRepo studentRepo;
+
 
     @GetMapping("/studentProfile")
     public String studentProfile(Model model, HttpServletRequest request) {
@@ -41,23 +44,31 @@ public class StudentController {
     public String saveProfile(@ModelAttribute Student formStudent, Model model) {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         EndUser endUser = userRepository.findByUsername(auth.getName());
+        Student student = studentRepo.findByUsername(auth.getName());
 
         if(endUser == null){
             // This is an error because the student is supposed to exist in DB
             model.addAttribute("view", "index");
             return "layout";
         }
-        else{
+        else if(student == null) {
             formStudent.setUsername(endUser.getUsername());
             formStudent.setPassword(endUser.getPassword());
             formStudent.setConfPassword(endUser.getConfPassword());
-
-            userRepository.save(formStudent);
+            formStudent.setHasProfile(true);
             userRepository.delete(endUser);
-
-            model.addAttribute("view", "index");
-            return "redirect:";
+            userRepository.save(formStudent);
+        } else {
+            student.setEmail(formStudent.getEmail());
+            student.setCv(formStudent.getCv());
+            student.setDiploma(formStudent.getDiploma());
+            student.setGradeAudit(formStudent.getGradeAudit());
+            student.setHasProfile(true);
+            userRepository.save(student);
         }
+
+        model.addAttribute("view", "index");
+        return "redirect:";
 
     }
